+++ src/MBrace.Azure/AssemblyInfo.fs
@@ -6,10 +6,10 @@ open System.Reflection
 [<assembly: AssemblyCompanyAttribute("Nessos Information Technologies")>]
 [<assembly: AssemblyCopyrightAttribute("© Nessos Information Technologies.")>]
 [<assembly: AssemblyTrademarkAttribute("MBrace")>]
-[<assembly: AssemblyMetadataAttribute("Release Signature","Version 0.10.0, Git Hash aa2eb9, Build Date 02092015 14:47 +03:00")>]
-[<assembly: AssemblyVersionAttribute("0.10.0")>]
-[<assembly: AssemblyFileVersionAttribute("0.10.0")>]
+[<assembly: AssemblyMetadataAttribute("Release Signature","Version 0.10.1, Git Hash 5de584, Build Date 07092015 15:59 +03:00")>]
+[<assembly: AssemblyVersionAttribute("0.10.1")>]
+[<assembly: AssemblyFileVersionAttribute("0.10.1")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "0.10.0"
+    let [<Literal>] Version = "0.10.1"